+++ models/src/main/java/com/vimeo/networking2/SurveyResponseChoice.kt
@@ -7,18 +7,22 @@ import com.vimeo.networking2.common.Entity
 /**
  * A survey response choice associated with a [SurveyQuestion].
  *
- * @param id A unique identifier for the response choice within the context of a [SurveyQuestion].
+ * @param analyticsId The id that should be used when logging this [SurveyResponseChoice].
+ * @param resourceKey A unique identifier for the response choice within the context of a [SurveyQuestion].
  * @param title A user-facing display title that represents the choice. This will be localized by the API.
  */
 @JsonClass(generateAdapter = true)
 data class SurveyResponseChoice(
 
-    @Json(name = "id")
-    val id: String? = null,
+    @Json(name = "analytics_id")
+    val analyticsId: String? = null,
+
+    @Json(name = "resource_key")
+    val resourceKey: String? = null,
 
     @Json(name = "title")
     val title: String? = null
 
 ) : Entity {
-    override val identifier: String? = id
+    override val identifier: String? = resourceKey
 }