+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Version 2.0.3 *(2021-08-10)*
+----------------------------
+- Improved support for user survey data by adding an `analyticsId` property to `SurveyQuestion`, `SurveyResponseChoice`, and `UserSegmentSurvey`. This property replaces any `id` properties the class originally had. 
+- Ensured that `SurveyResponseChoice` has a guid-based `resourceId` field that is tied to its `Entity.identifier`.
+- Renamed the `SurveyQuestion.emojiTitle` property to `titleEmoji`.
+- Added a `resourceKey` property to `SurveyResponseChoice`.
+
 Version 2.0.2 *(2021-08-06)*
 ----------------------------
 - Changed the `Play.source` property to the correct type, from `List<VideoSourceFile>` to `VideoSourceFile`.